backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
Pillow
requests
PyJWT
djangorestframework-simplejwt
